+++ desktop/cmp/form/FormField.js
@@ -46,6 +46,8 @@ export const [FormField, formField] = hoistCmp.withFactory({
 
     render({model, className, field, children, info, ...props}) {
 
+        console.log(props);
+
         // Resolve FieldModel
         const formContext = useContext(FormContext),
             formModel = formContext.model;
@@ -81,7 +83,9 @@ export const [FormField, formField] = hoistCmp.withFactory({
             labelAlign = defaultProp('labelAlign', props, formContext, 'left'),
             labelWidth = defaultProp('labelWidth', props, formContext, null),
             label = defaultProp('label', props, formContext, model?.displayName),
-            commitOnChange = defaultProp('commitOnChange', props, formContext, undefined);
+            commitOnChange = defaultProp('commitOnChange', props, formContext, undefined),
+            tooltipPosition = defaultProp('tooltipPosition', props, formContext, 'right'),
+            tooltipBoundary = defaultProp('tooltipBoundary', props, formContext, 'viewport');
 
         // Styles
         const classes = [childCssName];
@@ -110,7 +114,8 @@ export const [FormField, formField] = hoistCmp.withFactory({
                 target: childEl,
                 targetClassName: `xh-input ${displayNotValid ? 'xh-input-invalid' : ''}`,
                 targetTagName: !blockChildren.includes(getReactElementName(child)) || child.props.width ? 'span' : 'div',
-                position: 'right',
+                position: tooltipPosition,
+                boundary: tooltipBoundary,
                 disabled: !displayNotValid,
                 content: getErrorTooltipContent(errors)
             });
@@ -213,7 +218,21 @@ FormField.propTypes = {
     readonlyRenderer: PT.func,
 
     /** The indicator to display next to a required field. Defaults to `*`. */
-    requiredIndicator: PT.string
+    requiredIndicator: PT.string,
+
+    /**
+     * Minimal validation tooltip will try to fit within the corresponding boundary.
+     */
+    tooltipBoundary: PT.oneOf(['scrollParent', 'viewport', 'window']),
+
+    /** Position for minimal validation tooltip. */
+    tooltipPosition: PT.oneOf([
+        'top-left', 'top', 'top-right',
+        'right-top', 'right', 'right-bottom',
+        'bottom-right', 'bottom', 'bottom-left',
+        'left-bottom', 'left', 'left-top',
+        'auto', 'auto-start', 'auto-end'
+    ])
 };
 
 const readonlyChild = hoistCmp.factory({